Allow longer punish durations per data type

+++ src/main/java/de/korzhorz/punish/data/PunishReason.java
@@ -6,9 +6,9 @@ public class PunishReason {
     private int reasonId;
     private String reasonName;
     private PunishType punishType;
-    private int punishDuration;
+    private long punishDuration;
 
-    public PunishReason(int reasonId, String reasonName, PunishType punishType, int punishDuration) {
+    public PunishReason(int reasonId, String reasonName, PunishType punishType, long punishDuration) {
         this.reasonId = reasonId;
         this.reasonName = reasonName;
         this.punishType = punishType;
@@ -27,7 +27,7 @@ public class PunishReason {
         return punishType;
     }
 
-    public int getPunishDuration() {
+    public long getPunishDuration() {
         return punishDuration;
     }
 }
